feat(error-handling): add error handling in services
Add code that raise http exceptions to display details from an
error to the user

+++ services/tag.py
@@ -1,25 +1,49 @@
 from sqlalchemy.orm import Session
 from models.models import Tag as TagModel
 from schemas.tag import Tag
+from models.models import User as UserModel
+from fastapi import HTTPException, status
 
 
 class TagService():
+
     def __init__(self):
         pass
 
     async def get_tags(self, db: Session):
-        return db.query(TagModel).all()
+        tags = db.query(TagModel).all()
 
-    async def create_tag(self, tag: Tag, db: Session):
+        if not tags:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                                detail="No tags found")
+        return tags
+
+    async def create_tag(self, tag: Tag, db: Session, username: str):
+        user_exist = db.query(UserModel).filter(
+            UserModel.username == username).first()
+
+        if not user_exist:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                                detail="Sign in first")
+
         new_tag = TagModel(**tag.dict())
         db.add(new_tag)
         db.commit()
         return
 
-    async def delete_tag(self, id: int, db: Session):
+    async def delete_tag(self, id: int, db: Session, username: str):
+        user_exist = db.query(UserModel).filter(
+            UserModel.username == username).first()
+
+        if not user_exist:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                                detail="Sign in first")
+
         result = db.query(TagModel).filter(TagModel.id == id).first()
+
         if not result:
-            return
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                                detail="Tag not found")
         db.delete(result)
         db.commit()
         return
